+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Vz Keymap extension will be documented in this file.
 
+### [Unreleased]
+- WIP:
+  - Added undelete command (Ctrl+U) which restores deleted characters.
+
 ### [0.10.2] - 2020-08-02
 - Improved:
   - The behavior of Paste of a text which is from box-selection mode.